requirements.txt
pandas>=1.5
numpy>=1.23
statsmodels>=0.13
tensorflow-cpu>=2.10
